Implement PreferRepeatableRead
COMMENTS NEEDED: We may want to reconsider the interface the user uses
to specify it!
Currently it works by setting default_transaction_isolation to the
enum value PreferRepeatableRead.

If that is set, the parse and execute paths will strip it from the
state that is sent, and then potentially inject RepeatableRead as the
value when doing execute, if it is safe.