Implement PreferRepeatableRead, apply TransactionOptions to config state (#609)
Currently it is enabled by setting default_transaction_isolation or isolation in TransactionOptions
to the enum value PreferRepeatableRead.

If that is set, the parse and execute paths will strip it from the
state that is sent, and then potentially inject RepeatableRead as the
value when doing execute, if it is safe.