Initial implementation of schema reflection to Pydantic models
This adds a codegen mode `gel-generate models`, which introspects the
database schema and generates a full class layout for all types found in
it (including the builtin modules).

To execute reflection, run `gel-generate models` in a Gel project
folder.  Generated Python code will be placed into the `models/` package
in the current directory.

Generated Python classes are full Pydantic models and so support
serialization and introspection.