Initial implementation of autoreflected ORM (Pydantic version) (RFC 1042)
This adds a codegen mode `gel-generate models`, which introspects the
database schema and generates a full class layout for all types found in
it (including the builtin modules).

To execute reflection, run `gel-generate models` in a Gel project
folder.  Generated Python code will be placed into the `models/` package
in the current directory.

Generated Python classes are full Pydantic models and so support their
serialization and introspection.

In addition to that we implement a fully typesafe query builder that can
emit and compose EdgeQL expression with Python syntax, e.g:

    qexpr = default.Movie.select(title=True) \
        .filter(lambda m: m.release_year > 2000) \
        .order_by(title="asc") \
        .limit(10)

    data = client.query(qexpr)

The returned data is a list of Pydantic model instances that are capable
of tracking changed state and flushing the changes to the database:

    movie = default.Movie(
        title="New Movie",
        actors=[default.Person(name="me"],
    )

    client.save(movie)

See more at https://github.com/geldata/rfcs/blob/master/text/1042-fastapi.rst

Co-authored-by: Yury Selivanov <yury@geldata.com>