Add BuiltinUI class (#626)
To use the built-in UI you must:

1. Start a PKCE session
2. Redirect to the built-in UI URL with the PKCE challenge

This class wraps up that functionality in a simple interface.
Implementations will need to save the verifier in a cookie, and redirect
the user agent to the `BuiltinUIResponse.redirect_url`, and then
exchange the PKCE code and verifier for a token using the normal PKCE
code exchange.