models: Reflect non-magic operators as well
Add reflection of EdgeQL operators that have no corresponding Python
dunder magic.  These are reflected as overloaded functions in std with
names specified by the `std::identifier` annotation or the operator name
if its an identifier itself.  Operators that are clashing with Python
keywords (e.g `or`) are qualified with trailing underscore (`or_()`).

Fixes: #670