models: Declare properties as returning standard Python type _values_
The current property descriptors are typed as returning scalar type
reflections _both_ on class and instance attribute references.  This is
necessary for classes because of overloading, but values must be typed
as regular Python types (or subtypes thereof).

Initially I planned to construct an elaborate constraint system and
return something like `int[std.int64]` in order to allow such values to
be composed in the query builder without loss of casting fidelity, but
that turned out to be a bit too complex for the cause and I decided to
punt and simply return the base Python type instead.