models: Generate specializations for generic functions and operators
The std schema defines a number of functions and operators as fully
generic, e.g `distinct anytype -> anytype` or `range(anypoint)`.
Currently, we reflect those as generic functions as well.  The trouble
with that is the generic type variable can only bind to Gel type
reflections, not Python builtin types, so we are missing the ability to
use regular Python values with those callables.  We rectify that here
by synthesizing overloads from generics using the type inheritance
information.  This necessitate a bit of rework in callable reflection to
allow replacing parameter types in signatures.