Replicate uuid validation logic from Pydantic for GelModel.id
Unfortunately we can't just use pydantic facilities to validate
values for id fields as the field is frozen (users are not supposed
to override it) and we have a lot of other custom logic.

Another way to implement this would be to have a temporary model
to set an id to it, but the error message would show wrong model name
then, which is confusing.