Stop storing UNSET_UUID in models; don't use it as default for 'id'
Instead of initializing `.id` for new models to UNSET_UUID we now don't
set it at all. Instead of returning UNSET_UUID, new models will simply
throw an `AttributeError`. Which is ultimately a less surprising behavior
than getting a weird flavor of UUID type that they can't do anything
with in most cases (but would be pretty bad if it somehow surfaces in an
unexpected place).

We also no longer use it as a default for GelModel's `id` field. This
fixes warnings in pydantic, when it's failing to serialize UNSET_UUID
default value for rendering in JSON schemas.

GelModel gets a new internal attribute `__gel_new__`. When set to `True`
it meand the model hasn't yet been persisted. `__gel_commit__` flips it
to `False`.