models: Make all link descriptors symmetrical
Pyright (and, possibly, soon mypy) disable assignment type narrowing
when a descriptor is assymetrical (i.e its `__get__` is different from
`__set__`.  This is very unergonomic with respect to optional links, as
one has to assign to a variable and narrow that, e.g

    foo.optlink = non_empty_value
    assert foo.optlink is not None
    foo.optlink.prop  # still a type error

    temp = foo.optlink
    assert temp is not None
    temp.prop  # now works

This is quite bad and is the case because our `LinkWithProperties`
descriptors allow assigning the naked target type (via `_BMT_co`).

Do the less bad thing and narrow assignment to proxy model types.  This
now mandates `Foo.optlink.link(val)` when assigning.