models: Prevent erroneous use of boolean conversion in queries
Presently there is no way to overload boolean logic
operators in Python (`and`, `or`, `not`).  Likewise, `bool()` and `in`,
while overloadable via `__bool__` and `__contains__` require that the
return value is a `bool` so we cannot yield a query expression from
there.

At the same time, since most operators are overloaded, it is extremely
easy to unintentionally or unknowingly use implicit or explicit boolean
conversion on a query expression and get an obscure error.  In some
contexts invalid usage will be caught by the type checker, but where a
Python boolean value is accepted this would just be a silent bug.
Solve this by raising a `TypeError` from `__bool__` and `__contains__`
overloads on reflected types and expressions.  Go beyond just raising a
generic error and attempt to actually determine the context of boolean
conversion by looking at bytecode of the calling frame.

Fixes: #780