models: Ban `<expr> is [not] None`
Similarly to prohibiting `<expr> == None`, raise on `<expr> is None`
is used.  Unlike boolean ops, there is no idiomatic way of intercepting
identity checks, so we have to resort to AST rewrites, which limits the
scope of where the protection applies to lambdas in query builder
expreesssions.

Follow up to PR #799 (issue #780).