Fix ORDER BY in complex situations
Make IteratorStmt bind with WITH instead of FOR when
self_ref_must_bind.  This gives us a distinct name for the binding
but still relies on the Select or whatever to do the actual binding,
which makes ORDER BY work properly.

(Previously ORDER BY wouldn't work in these cases because it was
inside a FOR.)

Fixes #893.