Refetch computed for newly inserted objects (#869)
Given a schema
```
type A {
   total_val := sum(.<a[is B].val);
};
type B {
    a: A;
    val: int64;
};
```

If a user does:
```
a = default.A()
b = default.B(a=a, val=7)

client.sync(a, b)
```

The order of inserts will be `a`, then `b` and so the value of `a.total_val` fetched during the insert will be 0.

This PR adds a refetch query for the computed properties of newly inserted objects.

Tests are based on #848 and cover rewrites and triggers as well.