Fix problems with forward refs in the model (#906)
When `ModelFieldDescriptor.__get__` is invoked early, in the bowels of
both pydantic and our glue, it might fail to resolve the type because
the namespace isn't done being populated.

We were caching that, and it meant it wouldn't work later.
Don't cache it when it fails.

Also return a more obviously broken value from __get__ when resolution
fails, instead of `self`.