Prevent test discovery from loading models (#915)
Since we started putting models in tests/models, unittest (which
geltest uses also for this) test discovery will go and recursively
import models that are sitting around between test runs. This is rough
as some/all of them are about to be blown away and regenerated.

If we just switched from some branch making interesting changes, the
imports might fail.

unittest has a `load_tests` protocol that we can use to take charge of
our own test loading for a package... and then not do any.