Fix a bunch of issues with arrays as properties (#942)
Fix json_schema on arrays and multi properties.

Don't use DowncastingTrackedList (which I guess I will now remove in a
follow-up) because we don't want it's behavior of casting things to
types
like `std.str`.

Update the CoreSchemas of arrays, multiprops, and tuples to properly use
the underlying serializers -- for array and multi prop, we want to
convert
to a `list`, and then run a normal pydantic list serializer on the
output
type.

Without this, multi props of arrays wouldn't convert the interior arrays
into lists (leaving them as TrackedList), which caused problems like
breaking model_dump_json.


Fixes #937